+++ cs/CMakeLists.txt
@@ -252,6 +252,7 @@ set(OBJECT_SOURCES
   ${OBJECT_DIR}/class_object.c
   ${OBJECT_DIR}/deduplicate_key.c
   ${OBJECT_DIR}/elo.c
+  ${OBJECT_DIR}/identifier_store.cpp
   ${OBJECT_DIR}/lob_locator.cpp
   ${OBJECT_DIR}/object_accessor.c
   ${OBJECT_DIR}/object_description.cpp
@@ -276,7 +277,6 @@ set(OBJECT_SOURCES
   ${OBJECT_DIR}/work_space.c
   )
 set(OBJECT_HEADERS
-  ${OBJECT_DIR}/identifier_store.hpp
   ${OBJECT_DIR}/lob_locator.hpp
 )
