+++ cubrid/CMakeLists.txt
@@ -252,6 +252,7 @@ set(QUERY_HEADERS
 set(OBJECT_SOURCES
   ${OBJECT_DIR}/deduplicate_key.c
   ${OBJECT_DIR}/elo.c
+  ${OBJECT_DIR}/identifier_store.cpp
   ${OBJECT_DIR}/lob_locator.cpp
   ${OBJECT_DIR}/object_domain.c
   ${OBJECT_DIR}/object_primitive.c
@@ -261,7 +262,6 @@ set(OBJECT_SOURCES
   ${OBJECT_DIR}/transform.c
   )
 set(OBJECT_HEADERS
-  ${OBJECT_DIR}/identifier_store.hpp
   ${OBJECT_DIR}/lob_locator.hpp
 )
