+++ sa/CMakeLists.txt
@@ -290,6 +290,7 @@ set(OBJECT_SOURCES
   ${OBJECT_DIR}/class_object.c
   ${OBJECT_DIR}/deduplicate_key.c
   ${OBJECT_DIR}/elo.c
+  ${OBJECT_DIR}/identifier_store.cpp
   ${OBJECT_DIR}/lob_locator.cpp
   ${OBJECT_DIR}/object_description.cpp
   ${OBJECT_DIR}/object_domain.c
@@ -314,7 +315,6 @@ set(OBJECT_SOURCES
   ${OBJECT_DIR}/work_space.c
   )
 set(OBJECT_HEADERS
-  ${OBJECT_DIR}/identifier_store.hpp
   ${OBJECT_DIR}/lob_locator.hpp
 )
